third_party/WebKit/LayoutTests/fast/canvas-api/canvas-getImageData-invalid.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
test(function(t) {
    ctx = document.createElement('canvas').getContext('2d');

    try {
      ctx.getImageData(10, 10, 0, 10);
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }
    try {
      ctx.getImageData(10, 10, 10, 0);
    } catch (ex) {
      assert_equals(ex.name, "IndexSizeError");
    }
}, "Test the handling of invalid arguments in canvas getImageData().");
</script>
</body>
